+++ webui/react/src/pages/SignIn.tsx
@@ -3,8 +3,8 @@ import queryString from 'query-string';
 import React, { useEffect, useMemo, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 
-import LogoGoogle from 'assets/logo-sso-google-white.svg';
-import LogoOkta from 'assets/logo-sso-okta-white.svg';
+import LogoGoogle from 'assets/images/logo-sso-google-white.svg';
+import LogoOkta from 'assets/images/logo-sso-okta-white.svg';
 import AuthToken from 'components/AuthToken';
 import DeterminedAuth from 'components/DeterminedAuth';
 import Logo, { LogoType } from 'components/Logo';
