+++ webui/react/src/styles/fonts.scss
@@ -4,8 +4,8 @@
   font-style: normal;
   font-weight: normal;
   src:
-    url('fonts/objektiv-mk3-regular.woff2?07a67e') format('woff2'),
-    url('fonts/objektiv-mk3-regular.woff?07a67e') format('woff');
+    url('../assets/fonts/objektiv-mk3-regular.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/objektiv-mk3-regular.woff?07a67e') format('woff');
 }
 
 @font-face {
@@ -14,8 +14,8 @@
   font-style: italic;
   font-weight: normal;
   src:
-    url('fonts/objektiv-mk3-regular-italic.woff2?07a67e') format('woff2'),
-    url('fonts/objektiv-mk3-regular-italic.woff?07a67e') format('woff');
+    url('../assets/fonts/objektiv-mk3-regular-italic.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/objektiv-mk3-regular-italic.woff?07a67e') format('woff');
 }
 
 @font-face {
@@ -24,8 +24,8 @@
   font-style: normal;
   font-weight: bold;
   src:
-    url('fonts/objektiv-mk3-bold.woff2?07a67e') format('woff2'),
-    url('fonts/objektiv-mk3-bold.woff?07a67e') format('woff');
+    url('../assets/fonts/objektiv-mk3-bold.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/objektiv-mk3-bold.woff?07a67e') format('woff');
 }
 
 @font-face {
@@ -34,8 +34,8 @@
   font-style: italic;
   font-weight: bold;
   src:
-    url('fonts/objektiv-mk3-bold-italic.woff2?07a67e') format('woff2'),
-    url('fonts/objektiv-mk3-bold-italic.woff?07a67e') format('woff');
+    url('../assets/fonts/objektiv-mk3-bold-italic.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/objektiv-mk3-bold-italic.woff?07a67e') format('woff');
 }
 
 @font-face {
@@ -44,8 +44,8 @@
   font-style: normal;
   font-weight: bold;
   src:
-    url('fonts/source-code-pro-bold.woff2?07a67e') format('woff2'),
-    url('fonts/source-code-pro-bold.woff?07a67e') format('woff');
+    url('../assets/fonts/source-code-pro-bold.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/source-code-pro-bold.woff?07a67e') format('woff');
 }
 
 @font-face {
@@ -54,6 +54,6 @@
   font-style: normal;
   font-weight: normal;
   src:
-    url('fonts/source-code-pro-regular.woff2?07a67e') format('woff2'),
-    url('fonts/source-code-pro-regular.woff?07a67e') format('woff');
+    url('../assets/fonts/source-code-pro-regular.woff2?07a67e') format('woff2'),
+    url('../assets/fonts/source-code-pro-regular.woff?07a67e') format('woff');
 }
