+++ webui/react/src/styles/icons.scss
@@ -4,8 +4,8 @@
   font-style: normal;
   font-weight: normal;
   src:
-    url('fonts/determined-ai.woff2?1v4ahh') format('woff2'),
-    url('fonts/determined-ai.woff?1v4ahh') format('woff');
+    url('../assets/fonts/determined-ai.woff2?1v4ahh') format('woff2'),
+    url('../assets/fonts/determined-ai.woff?1v4ahh') format('woff');
 }
 
 [class^='icon-'],
